+++ Changelog.md
@@ -3,6 +3,7 @@
 - Functions return to default when loco dropped
 - DCC address labels changed from 'CV' to 'DCC Addr' or 'Addr'
 - added setting to turn off the timestamp in the log
+- improvement to when it thinks the CS is ready to accept commands
 
 # Version 1.3.9
 - Improvements for when a separate throttle is controlling the same loco